Move get_valid_axis_keys
